# Conflicts:
#	Assets/Resources/Textures/Sprites/UnitIcon.png
#	Assets/Resources/Textures/Sprites/UnitIcon.png.meta
#	Assets/Scripts/Controls/MenuController.cs

+++ Assets/Scripts/Entities/Interfaces/IRepairable.cs
@@ -2,6 +2,6 @@
 public interface IRepairable
 {
     bool NeedsRepairing();
-    void Repair(int amount);
+    void Repair(float amount);
     void FullRepair();
 }
